requirements.txt
python-dotenv==1.0.0
numpy
pandas
backoff==2.2.1
google-cloud-aiplatform
google-cloud-aiplatform[langchain,reasoningengine]
langchain-google-community
google-cloud-discoveryengine
google-api-python-client
langchain
pypdf
langchain-google-vertexai
langchain-chroma
langgraph
chromadb
tenacity
chainlit
flashrank
